fix #531: change uncompressed chunk `size` and `num_samples` from short to `size_t` and `unsigned int`

+++ src/chunk.h
@@ -13,8 +13,8 @@
 typedef struct Chunk {
     timestamp_t base_timestamp;
     Sample *samples;
-    short num_samples;
-    short size;
+    unsigned int num_samples;
+    size_t size;
 } Chunk;
 
 typedef struct ChunkIterator {
